Inconsistencies with job_ids fixed
In run.py:

In the function run() in run.py, the ids should be provided as a list of
the form range(num_of_runs)
to create result-files beginning with 001-results.json.
As the list of ids should begin with 0, but the list of finished_ids
created by the get_finished_ids() function
is shifted by +1, the id filtering for the option force_rerun == False
is not working correctly

In hypersearch.py:

range(1, trials_per_point + 1) causes the enumeration of the
results.json files to be shifted
about + 1, as the job_id is incremented by 1 in the _run_helper function
within run.py

This leads to problems if the hyperparameter search is resumed.